+++ udata_front/theme/gouvfr/assets/js/components/discussions/Threads.vue
@@ -40,14 +40,17 @@
           <Well type="secondary" color="success" class="fr-mt-2w" data-cy="threadFromURL">
             <div class="fr-grid-row fr-grid-row--middle justify-between">
               {{
-                $t("You are seeing a specific discussion about this dataset")
+                $t("You are seeing a specific discussion about this dataset.")
               }}
               <button class="fr-btn--close fr-btn fr-mr-0" @click.prevent="resetHash">
                 {{$t('Close')}}
               </button>
             </div>
           </Well>
-          <thread v-bind="threadFromURL"></thread>
+          <thread v-if="threadFromURL" v-bind="threadFromURL"></thread>
+          <Well v-else type="secondary" color="warning" class="fr-mt-2w">
+            {{ $t("This discussion doesn't exist or was deleted.") }}
+          </Well>
           <button
             class="nav-link nav-link--no-icon text-decoration-none fr-link fr-mt-9v fr-link--icon-left fr-icon-arrow-right-s-line"
             @click.prevent="resetHash"
@@ -117,8 +120,8 @@ export default defineComponent({
     },
   },
   setup(props) {
-    const {t} = useI18n();
-    const toast = useToast();
+    const { t } = useI18n();
+    const { toast } = useToast();
 
     const discussions = ref([]);
 
@@ -134,7 +137,7 @@ export default defineComponent({
      * @property {String} title
      * @property {String} url
      * @property {String} closed - date when the thread was closed, as string
-     * @property {import("../../types").User} closed_by - User who closed the thread
+     * @property {import("@etalab/data.gouv.fr-components").User} closed_by - User who closed the thread
      */
 
     /** @type {import("vue").Ref<Thread | null>} */