+++ udata_front/theme/gouvfr/assets/js/pages/PublishingForm/Step1PublishingType.vue
@@ -97,8 +97,8 @@ import Stepper from '../../components/Form/Stepper/Stepper.vue';
 import Container from '../../components/Ui/Container/Container.vue';
 import Well from '../../components/Ui/Well/Well.vue';
 import { api_doc_external_link, guides_harvesting_url, schema_publishing_url, support_url, catalog_url, demo_server_name, demo_server_url, title } from "../../config";
-import datasetIcon from "svg/illustrations/dataset.svg";
-import schemaIcon from "svg/illustrations/schema.svg";
+import datasetIcon from "../../../../templates/svg/illustrations/dataset.svg";
+import schemaIcon from "../../../../templates/svg/illustrations/schema.svg";
 
 export default defineComponent({
   components: { ActionCard, Container, Stepper, Well },