+++ udata_front/theme/gouvfr/datagouv-components/src/types/users.ts
@@ -1,0 +1,10 @@
+export type User = {
+  id: string;
+  slug?: string;
+  first_name: string;
+  last_name: string;
+  avatar?: string | null;
+  page: string;
+  avatar_thumbnail?: string | null;
+  roles?: Array<string>;
+}